Update README.md
In the left segment, we have available dishes that can be selected according to the order. On the right side, we have a display that shows completed orders and an empty field where the invoice is displayed when the Pay button is clicked. On the lower right, the program shows the fields Sub Total, Total amount to be paid, Cash field where the amount of the customer's payment is entered into the empty field, Tax, which can be changed arbitrarily, as well as the amount of discount and Balance, which is calculated on the basis of the entered amount of Cash.
The program also offers the following options, Delete Row deletes the lines of executed orders, then we have the Delete Bill button, when pressed it deletes the bill on the display. The Delete Cash button deletes the manually entered cash payment amount field. Pressing the NEW button clears all fields and enables the execution of a new order.